Simplify JDBC result set iterator implementation
