+++ client/trino-jdbc/src/main/java/io/trino/jdbc/CancellableIterator.java
@@ -1,0 +1,45 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.trino.jdbc;
+
+import java.util.Iterator;
+
+interface CancellableIterator<T>
+        extends Iterator<T>
+{
+    void cancel();
+
+    static <T> CancellableIterator<T> wrap(Iterator<T> iterator)
+    {
+        return new CancellableIterator<T>() {
+            @Override
+            public void cancel()
+            {
+                // noop
+            }
+
+            @Override
+            public boolean hasNext()
+            {
+                return iterator.hasNext();
+            }
+
+            @Override
+            public T next()
+            {
+                return iterator.next();
+            }
+        };
+    }
+}
